Change color of Correlation plot IDS preference S1 and S2
Changing the color of method in the Correlation between IDS Preference in Session 1 and Session 2 plot to match plots in Supplement, which match colors of method in original MB1 paper
Note: There's an a in front of each method within the plot. Not sure where this comes from. Maybe this can be changed to a dot/line ...